cuda/pyproject.toml
[tool.poetry]
name = "manager.py"
version = "0.1.0"
description = "CUDA Docker CI Scripts"
authors = ["Jesus Alvarez <jesusa@nvidia.com>"]

[tool.poetry.dependencies]
python = ">3.8,<4"
pyyaml = ">6.0"
plumbum = "*"
jinja2 = "*"
coloredlogs = "*"
glom = "*"
docker = "*"
deepdiff = "^4.0"
mmh3 = "^2.5"
requests = "^2.22"
pyyaml-include = "^1.2"
retry = "^0.9.2"
packaging = "^20.9"
beeprint = "^2.4.10"
six = "^1.16.0"
urllib3 = "1.26.9"

[tool.poetry.dev-dependencies]
pytest = ">3.6"
mypy = "^0.950"
types-requests = "^2.27.25"
types-retry = "^0.9.7"
types-PyYAML = "^6.0.7"
setuptools = "^65.3.0"
wheel = "0.38.4"

[tool.pytest.ini_options]
filterwarnings = [
    'ignore:.*distutils.*:DeprecationWarning',
]

[tool.black]
line-length = 90
